requirements.txt
langchain
langchain-community
ollama
faiss-cpu
tree-sitter
GitPython
tqdm
